# (chore): optimize string writes