Optimize string writes by using zero copy string to byte conversion via unsafe package.

```
goos: darwin
goarch: arm64
pkg: github.com/johannes94/hashstructure
cpu: Apple M4 Pro
                  │   old.txt    │               new.txt                │
                  │    sec/op    │    sec/op     vs base                │
Map-14              3.057µ ± 18%   1.720µ ±  7%  -43.74% (p=0.000 n=10)
String/default-14   60.32n ± 17%   35.44n ±  9%  -41.24% (p=0.000 n=10)
String/xxhash-14    31.06n ± 22%   24.14n ± 20%  -22.26% (p=0.002 n=10)
geomean             178.9n         113.7n        -36.42%

                  │  old.txt   │               new.txt                │
                  │    B/op    │    B/op     vs base                  │
Map-14              985.0 ± 0%   633.0 ± 0%  -35.74% (p=0.000 n=10)
String/default-14   56.00 ± 0%   24.00 ± 0%  -57.14% (p=0.000 n=10)
String/xxhash-14    16.00 ± 0%   16.00 ± 0%        ~ (p=1.000 n=10) ¹
geomean             95.92        62.41       -34.94%
¹ all samples are equal

                  │  old.txt   │               new.txt                │
                  │ allocs/op  │ allocs/op   vs base                  │
Map-14              69.00 ± 0%   39.00 ± 0%  -43.48% (p=0.000 n=10)
String/default-14   3.000 ± 0%   2.000 ± 0%  -33.33% (p=0.000 n=10)
String/xxhash-14    1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean             5.915        4.273       -27.77%
¹ all samples are equal
```